+++ src/pygaps/calculations/models_isotherm/henry.py
@@ -121,10 +121,10 @@ class Henry(IsothermBaseModel):
 
         Parameters
         ----------
-        loading_key : str
-            Loading data.
-        pressure_key : str
+        pressure : ndarray
             Pressure data.
+        loading : ndarray
+            Loading data.
 
         Returns
         -------