+++ src/pygaps/calculations/models_isotherm/langmuir.py
@@ -161,10 +161,10 @@ class Langmuir(IsothermBaseModel):
 
         Parameters
         ----------
-        loading_key : str
-            Loading data.
-        pressure_key : str
+        pressure : ndarray
             Pressure data.
+        loading : ndarray
+            Loading data.
 
         Returns
         -------