+++ src/pygaps/calculations/models_isotherm/tslangmuir.py
@@ -144,10 +144,10 @@ class TSLangmuir(IsothermBaseModel):
 
         Parameters
         ----------
-        loading_key : str
-            Loading data.
-        pressure_key : str
+        pressure : ndarray
             Pressure data.
+        loading : ndarray
+            Loading data.
 
         Returns
         -------